# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/clinical-reference/{domain}": {
      "get": {
        "description": "domain ∈ clinical-procedures | bedside-interpretation | preventive-care | growth-development. Filters: category, q. Returns summaries; fetch full sections by slug.",
        "operationId": "ClinicalReferenceController_list",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "clinical-procedures",
                "bedside-interpretation",
                "preventive-care",
                "growth-development"
              ],
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List reference cards in a clinical-reference domain",
        "tags": [
          "clinical-reference"
        ]
      }
    },
    "/api/v1/clinical-reference/{domain}/{slug}": {
      "get": {
        "description": "Returns the full card: summary, titled sections (bullets or ordered steps), red flags and citations. Reference content only — no patient identity is processed.",
        "operationId": "ClinicalReferenceController_get",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "clinical-procedures",
                "bedside-interpretation",
                "preventive-care",
                "growth-development"
              ],
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a reference card by domain + slug",
        "tags": [
          "clinical-reference"
        ]
      }
    },
...
    "/api/v1/clinical-audits": {
      "get": {
        "operationId": "ClinicalAuditsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List clinical audits for the calling integrator",
        "tags": [
          "clinical-audits"
        ]
      },
      "post": {
        "description": "Binds a named query (connectionId + namedQueryId) to a set of custom-rule IDs, optional policy IDs to cite, a threshold (absolute count or percentage), and the notification-channel IDs to dispatch on breach.",
        "operationId": "ClinicalAuditsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a clinical audit",
        "tags": [
          "clinical-audits"
        ]
      }
    },
    "/api/v1/clinical-audits/{id}": {
      "get": {
        "operationId": "ClinicalAuditsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single audit definition by id",
        "tags": [
          "clinical-audits"
        ]
      },
      "patch": {
        "operationId": "ClinicalAuditsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update audit name / rules / policies / threshold / channels",
        "tags": [
          "clinical-audits"
        ]
      },
      "delete": {
        "operationId": "ClinicalAuditsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an audit",
        "tags": [
          "clinical-audits"
        ]
      }
    },
    "/api/v1/clinical-audits/{id}/run": {
      "post": {
        "description": "Executes the configured named query against the integrator-registered connection, evaluates each row against the configured custom rules, cites uploaded policies, computes threshold breach, and dispatches alerts via configured channels when the threshold trips. Returns full structured result (row-level verdicts use anonymous row indices — never patient identifiers).",
        "operationId": "ClinicalAuditsController_run",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run an audit on demand",
        "tags": [
          "clinical-audits"
        ]
      }
    },
...
    "/cds-services/{serviceId}/feedback": {
      "post": {
        "description": "Accepts an array of feedback entries per CDS Hooks 1.0 §6. Each entry references a card UUID we minted in a prior /cds-services response. PHI-free — only the outcome + override-reason code + optional comment is persisted.",
        "operationId": "CdsFeedbackSpecController_ingest",
        "parameters": [
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "CDS Hooks 1.0 feedback ingest",
        "tags": [
          "cds-hooks"
        ]
      }
    },
    "/api/v1/cds-feedback/summary": {
      "get": {
        "description": "Returns total feedback, accepted, overridden, override-rate %, and the top 3 override reason codes for every rule that has feedback. Sorted by total feedback descending. Rules being overridden ≥ 50 % of the time are flagged for editorial review (use this signal to retire false-positive rules).",
        "operationId": "CdsFeedbackOperatorController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Per-rule feedback aggregation",
        "tags": [
          "cds-feedback"
        ]
      }
    },
    "/api/v1/cds-feedback": {
      "get": {
        "operationId": "CdsFeedbackOperatorController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List feedback entries (latest first) optionally filtered by ruleId",
        "tags": [
          "cds-feedback"
        ]
      }
    },
...
  },
  "info": {
    "title": "VedaMD API",
    "description": "Source-available, stateless Clinical Decision Support API. CDS Hooks + REST/JSON. VedaMD never stores patient data — records remain with your EMR.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}